+++ internal/api/openai/v1/types.go
@@ -1,4 +1,4 @@
-package v1
+package openai
 
 import (
 	"encoding/json"
