+++ internal/backend/ollama/convert.go
@@ -2,7 +2,7 @@ package ollama
 
 import (
 	ollama "github.com/danilofalcao/cursor-deepseek/internal/api/ollama/v1"
-	openai "github.com/danilofalcao/cursor-deepseek/internal/api/openai/v1"
+	"github.com/danilofalcao/cursor-deepseek/internal/api/openai/v1"
 )
 
 func convertMessages(messages []openai.Message) []ollama.Message {
