+++ internal/backend/openrouter/utils.go
@@ -1,0 +1,56 @@
+package openrouter
+
+import (
+	"compress/flate"
+	"compress/gzip"
+	"fmt"
+	"io"
+	"net/http"
+
+	"github.com/andybalholm/brotli"
+)
+
+func copyHeaders(dst, src http.Header) {
+	// Headers to skip
+	skipHeaders := map[string]bool{
+		"Content-Length":    true,
+		"Content-Encoding":  true,
+		"Transfer-Encoding": true,
+		"Connection":        true,
+	}
+
+	for k, vv := range src {
+		if !skipHeaders[k] {
+			for _, v := range vv {
+				dst.Add(k, v)
+			}
+		}
+	}
+}
+
+func readResponse(resp *http.Response) ([]byte, error) {
+	var reader io.Reader = resp.Body
+
+	switch resp.Header.Get("Content-Encoding") {
+	case "gzip":
+		gzReader, err := gzip.NewReader(resp.Body)
+		if err != nil {
+			return nil, fmt.Errorf("error creating gzip reader: %v", err)
+		}
+		defer gzReader.Close()
+		reader = gzReader
+	case "br":
+		reader = brotli.NewReader(resp.Body)
+	case "deflate":
+		reader = flate.NewReader(resp.Body)
+	}
+
+	return io.ReadAll(reader)
+}
+
+func truncateString(s string, maxLen int) string {
+	if len(s) <= maxLen {
+		return s
+	}
+	return s[:maxLen] + "..."
+}
